new controller, break down ChronicleService into several beans
