+++ src/main/java/com/openlattice/chronicle/services/upload/AppDataUploadManager.java
@@ -1,0 +1,20 @@
+package com.openlattice.chronicle.services.upload;
+
+import com.google.common.collect.SetMultimap;
+
+import java.util.List;
+import java.util.UUID;
+
+/**
+ * @author alfoncenzioka &lt;alfonce@openlattice.com&gt;
+ */
+public interface AppDataUploadManager {
+
+    Integer logData(
+            UUID organizationId,
+            UUID studyId,
+            String participantId,
+            String datasourceId,
+            List<SetMultimap<UUID, Object>> data
+    );
+}
