+++ src/main/java/com/openlattice/chronicle/services/v2/ChronicleServiceV2.java
@@ -1,23 +1,4 @@
-/*
- * Copyright (C) 2018. OpenLattice, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * You can contact the owner of the copyright at support@openlattice.com
- */
-
-package com.openlattice.chronicle.services;
+package com.openlattice.chronicle.services.v2;
 
 import com.google.common.base.Optional;
 import com.google.common.collect.SetMultimap;
@@ -34,16 +15,22 @@ import java.util.Set;
 import java.util.UUID;
 
 /**
- * @author Matthew Tamayo-Rios &lt;matthew@openlattice.com&gt;
+ * @author alfoncenzioka &lt;alfonce@openlattice.com&gt;
  */
-public interface ChronicleService {
-    //  TODO: add in throws exception!
+public interface ChronicleServiceV2 {
+    Integer logData(
+            UUID studyId,
+            String participantId,
+            String datasourceId,
+            List<SetMultimap<UUID, Object>> data );
+
     Integer logData(
             UUID organizationId,
             UUID studyId,
             String participantId,
             String datasourceId,
-            List<SetMultimap<UUID, Object>> data );
+            List<SetMultimap<UUID, Object>> data
+    );
 
     UUID registerDatasource( UUID organizationId, UUID studyId, String participantId, String datasourceId, Optional<Datasource> datasource );
 
@@ -98,6 +85,6 @@ public interface ChronicleService {
     ChronicleQuestionnaire getQuestionnaire( UUID organizationId, UUID studyId, UUID questionnaireEKID );
 
     Map<UUID, Map<FullQualifiedName, Set<Object>>> getStudyQuestionnaires( UUID organizationId, UUID studyId );
-    
+
     void submitQuestionnaire( UUID organizationId, UUID studyId, String participantId, Map<UUID, Map<FullQualifiedName, Set<Object>>> questionnaireResponses );
 }
